requirements.txt
streamlit>=1.35.0
requests>=2.31.0
openpyxl>=3.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
